requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.6.0
openai>=1.54.0
anthropic>=0.39.0
google-genai>=0.3.0
chromadb>=0.5.20
bcrypt>=4.2.0
python-frontmatter>=1.1.0
beautifulsoup4>=4.12.3
httpx>=0.27.2
python-slugify>=8.0.4
